refactor: 패키지 구조 리팩토링
- 도메인 클래스의 패키지를 역할에 따라 분리
