--- src/main/java/response/handler/Handler.java
@@ -1,10 +1,0 @@
-package response.handler;
-
-import request.Request;
-import response.Response;
-
-import java.io.IOException;
-
-public interface Handler {
-    Response handle(Request request) throws IOException;
-}
