+++ src/main/java/webserver/http/common/StatusCode.java
@@ -1,4 +1,4 @@
-package common;
+package webserver.http.common;
 
 public enum StatusCode {
