+++ src/main/java/webserver/http/response/Response.java
@@ -1,4 +1,4 @@
-package response;
+package webserver.http.response;
 
 public class Response {
